backend/pyproject.toml
[project]
name = "arctic-ice-monitoring"
version = "0.1.0"
description = "Arctic Ice Monitoring Platform with ML-powered predictions and real-time satellite data"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Arctic Ice Monitoring Team"}
]

dependencies = [
    # Core dependencies shared across backend and ML
    "numpy>=1.26.0,<2.0.0",
    "pandas==2.1.3",
    "Pillow==10.1.0",
    "requests==2.31.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "tqdm==4.66.1",
    "rasterio>=1.3.9",
    "shapely>=2.0.2",
    "apscheduler>=3.11.2",
    "pyproj>=3.6.1",
]

[project.optional-dependencies]
backend = [
    # FastAPI and server
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",

    # Database
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    "psycopg2-binary==2.9.9",
    "geoalchemy2==0.14.2",

    # Redis and Celery
    "redis==5.0.1",
    "celery==5.3.4",

    # Machine Learning
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    # "tensorflow==2.14.0", # Removing due to install issues
    "scikit-learn==1.3.2",

    # Geospatial processing
    "rasterio==1.3.9",
    "shapely==2.0.2",
    "geojson==3.1.0",

    # Satellite data
    "sentinelsat==1.2.1",

    # Authentication
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",

    # Configuration
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "email-validator",
    "global-land-mask",
]

ml = [
    # ML Training
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    # "tensorflow==2.14.0",
    "scikit-learn==1.3.2",

    # Visualization
    "matplotlib==3.8.2",
    "seaborn==0.13.0",

    # Satellite data processing
    "rasterio==1.3.9",
    "sentinelsat==1.2.1",
    "netCDF4==1.6.5",
    "h5py==3.10.0",
    "xarray==2023.12.0",
]

dev = [
    # Development tools
    "jupyter==1.0.0",
    "jupyterlab==4.0.9",
    "ipykernel==6.27.1",
    "ipywidgets==8.1.1",

    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "httpx==0.25.2",

    # Code quality
    "black==23.12.0",
    "ruff==0.1.8",
    "mypy==1.7.1",
]

# Install everything
all = [
    "arctic-ice-monitoring[backend,ml,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []



[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | env
  | _build
  | build
  | dist
  | node_modules
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (black handles this)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=backend",
    "--cov=ml-models",
    "--cov-report=term-missing",
    "--cov-report=html",
]
